Sync: 115. Distinct Subsequences (Hard) - Leet2Git

+++ 115_Distinct_Subsequences.cpp
@@ -21,7 +21,7 @@ public:
         int m = s.size();
         int n = t.size();
 
-        vector<vector<int>>  dp(m + 1, vector<int>(n + 1, 0));
+        vector<vector<unsigned long long>>  dp(m + 1, vector<unsigned long long>(n + 1, 0));
     
         for(int i = 0; i <= m; i++) {
             dp[i][n] = 1;
@@ -38,6 +38,6 @@ public:
             }
         }
 
-        return dp[0][0];
+        return (int)dp[0][0];
     }
 };
